# fix performance regression from PR #109