The PR switches to hash data structure from array for storing supported metric types per implementation. I don't know the details of hash table implementation in Ruby but considering there are not many entries lookup should be O(1) versus O(n) when it was array.

I have not done the actual benchmarking yet - so it's only theory at this point. I'll wait for @fmejia97 to publish his benchmarking script first.

/cc @Shopify/edgescale