@fmejia97 2018-05-22
Benchmarking results:

**`TLDR: Performance improved on average by 6.22%`**

**`New changes`**:
``` Warming up --------------------------------------
increment metric benchmark
                        18.315k i/100ms
measure metric benchmark
                        14.632k i/100ms
gauge metric benchmark
                        17.064k i/100ms
set metric benchmark    15.022k i/100ms
event metric benchmark
                        18.881k i/100ms
service check metric benchmark
                        18.351k i/100ms
Calculating -------------------------------------
increment metric benchmark
                        198.699k (± 5.3%) i/s -      1.007M in   5.083605s
measure metric benchmark
                        148.297k (±11.9%) i/s -    731.600k in   5.030920s
gauge metric benchmark
                        209.135k (± 8.1%) i/s -      1.041M in   5.018486s
set metric benchmark    161.045k (± 8.8%) i/s -    811.188k in   5.079304s
event metric benchmark
                        198.675k (± 6.5%) i/s -      1.001M in   5.059252s
service check metric benchmark
                        205.463k (± 9.6%) i/s -      1.028M in   5.055052s
Run options: --seed 62606 
```

**``` current master: ```**
```
Warming up --------------------------------------
increment metric benchmark
                        16.798k i/100ms
measure metric benchmark
                        13.989k i/100ms
gauge metric benchmark
                        18.051k i/100ms
set metric benchmark    13.910k i/100ms
event metric benchmark
                        16.709k i/100ms
service check metric benchmark
                        17.359k i/100ms
Calculating -------------------------------------
increment metric benchmark
                        186.545k (± 9.8%) i/s -    923.890k in   5.010993s
measure metric benchmark
                        147.626k (± 9.2%) i/s -    741.417k in   5.084888s
gauge metric benchmark
                        196.508k (± 5.4%) i/s -    992.805k in   5.068009s
set metric benchmark    155.675k (± 9.0%) i/s -    778.960k in   5.055491s
event metric benchmark
                        193.414k (± 7.9%) i/s -    969.122k in   5.054284s
service check metric benchmark
                        188.934k (± 9.6%) i/s -    937.386k in   5.036664s
Run options: --seed 38544
```